fix: resolve admin ci failures
