+++ admin-ui/src/styles.css
@@ -2967,181 +2967,3 @@ pre.empty {
   color: var(--primary);
   font-weight: 600;
 }
-
-/* TokenTracker-style hero metrics — large display numbers for headline stats. */
-.stats-hero {
-  display: grid;
-  gap: 1rem;
-  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-  margin-bottom: 1.25rem;
-}
-
-.hero-metric {
-  background: var(--card);
-  border: 1px solid var(--border);
-  border-radius: var(--radius-lg);
-  box-shadow: var(--shadow-sm);
-  display: grid;
-  gap: 0.35rem;
-  padding: 1.3rem 1.4rem 1.4rem;
-  position: relative;
-  overflow: hidden;
-  transition: border-color 0.2s ease, box-shadow 0.2s ease, transform 0.2s ease;
-}
-
-.hero-metric::before {
-  content: '';
-  position: absolute;
-  inset: 0 0 auto 0;
-  height: 3px;
-  background: linear-gradient(90deg, var(--brand-light), var(--brand));
-  opacity: 0;
-  transition: opacity 0.2s ease;
-}
-
-.hero-metric:hover {
-  border-color: var(--accent-line);
-  box-shadow: var(--shadow-md);
-  transform: translateY(-1px);
-}
-
-.hero-metric:hover::before,
-.hero-metric.accent::before {
-  opacity: 1;
-}
-
-.hero-metric .hero-label {
-  color: var(--muted-foreground);
-  font-size: 0.7rem;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-}
-
-.hero-metric .hero-value {
-  color: var(--foreground);
-  font-family: var(--font-sans);
-  font-size: clamp(2rem, 4.4vw, 3.25rem);
-  font-weight: 700;
-  letter-spacing: -0.02em;
-  line-height: 1.02;
-  font-variant-numeric: tabular-nums;
-}
-
-.hero-metric.accent .hero-value {
-  color: var(--brand-dark);
-}
-
-.hero-metric .hero-detail {
-  color: var(--muted-foreground);
-  font-size: 0.78rem;
-  line-height: 1.4;
-}
-
-.hero-metric .hero-detail strong {
-  color: var(--brand-dark);
-  font-weight: 700;
-}
-
-/* Stats charts (CSS-only) */
-.stats-charts {
-  display: grid;
-  gap: 1.25rem;
-  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
-  margin-bottom: 1rem;
-}
-
-.chart-card {
-  background: var(--card);
-  border: 1px solid var(--border);
-  border-radius: var(--radius-lg);
-  min-height: 12rem;
-  padding: 1rem 1.1rem;
-}
-
-.chart-title {
-  color: var(--foreground);
-  font-family: var(--font-sans);
-  font-size: 0.75rem;
-  font-weight: 600;
-  letter-spacing: 0.06em;
-  margin-bottom: 0.65rem;
-  text-transform: uppercase;
-}
-
-.hbar-row {
-  align-items: center;
-  display: grid;
-  gap: 0.5rem;
-  grid-template-columns: minmax(0, 1fr) minmax(0, 2.2fr) 2.5rem;
-  margin-bottom: 0.4rem;
-}
-
-.hbar-label {
-  color: var(--foreground);
-  font-family: var(--font-mono);
-  font-size: 0.6875rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.hbar-track {
-  background: var(--muted);
-  border-radius: 3px;
-  height: 0.55rem;
-  overflow: hidden;
-}
-
-.hbar-fill {
-  background: linear-gradient(90deg, var(--primary), var(--primary));
-  border-radius: 3px;
-  height: 100%;
-  min-width: 2px;
-  transition: width 0.25s ease;
-}
-
-.hbar-count {
-  color: var(--foreground);
-  font-family: var(--font-mono);
-  font-size: 0.6875rem;
-  font-weight: 600;
-  text-align: right;
-}
-
-.hourly-chart {
-  align-items: flex-end;
-  display: flex;
-  gap: 2px;
-  height: 120px;
-  margin-top: 0.25rem;
-  padding-top: 0.25rem;
-}
-
-.hour-col {
-  align-items: center;
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  height: 100%;
-  justify-content: flex-end;
-  min-width: 0;
-}
-
-.hour-bar {
-  background: linear-gradient(180deg, var(--primary), var(--ok));
-  border-radius: 2px 2px 0 0;
-  min-height: 2px;
-  transition: height 0.25s ease;
-  width: 100%;
-}
-
-.hour-tick {
-  color: var(--muted-foreground);
-  font-family: var(--font-mono);
-  font-size: 0.5625rem;
-  margin-top: 2px;
-  min-height: 0.75rem;
-  text-align: center;
-  width: 100%;
-}
