+++ crates/dcc-mcp-telemetry/Cargo.toml
@@ -28,6 +28,7 @@ tracing-opentelemetry = { version = "0.33" }
 
 # OTLP exporter (optional — only for production export to Jaeger/Tempo/etc.)
 opentelemetry-otlp = { version = "0.32", features = ["grpc-tonic", "trace", "metrics"], optional = true }
+tonic = { version = "0.14", optional = true }
 
 # Prometheus exporter (optional — enabled via the `prometheus` feature).
 # Zero-cost when disabled; no code paths are compiled into the wheel.
@@ -47,7 +48,7 @@ python-bindings = ["pyo3"]
 # Enable `pyo3-stub-gen` annotations on this crate's PyO3 types.
 stub-gen = ["python-bindings", "dep:pyo3-stub-gen", "dep:pyo3-stub-gen-derive"]
 # Enable OTLP export to Jaeger / Grafana Tempo / Prometheus
-otlp-exporter = ["opentelemetry-otlp"]
+otlp-exporter = ["opentelemetry-otlp", "dep:tonic"]
 # Enable the Prometheus text-exposition exporter (issue #331).
 # Surfaces a `PrometheusExporter` type that reads from `ToolRecorder`
 # state and can be rendered to a Prometheus scrape body. The HTTP crate
