+++ tests/test_gateway_guardian.py
@@ -451,9 +451,10 @@ def test_guardian_run_catches_crash_and_increments_crash_count(monkeypatch):
 
     guardian.start()
     try:
-        assert _wait_until(lambda: crash_reported.is_set() or guardian.status().get("crash_count", 0) >= 1), (
-            "Expected guardian crash status to be published"
-        )
+        assert _wait_until(
+            lambda: crash_reported.is_set() or guardian.status().get("crash_count", 0) >= 1,
+            timeout=30.0,
+        ), "Expected guardian crash status to be published"
     finally:
         guardian.stop(timeout=2.0)
 
@@ -494,9 +495,10 @@ def test_guardian_run_continues_after_exception(monkeypatch):
 
     guardian.start()
     try:
-        assert _wait_until(lambda: crash_reported.is_set() or guardian.status().get("crash_count", 0) >= 1), (
-            "Expected guardian crash status to be published"
-        )
+        assert _wait_until(
+            lambda: crash_reported.is_set() or guardian.status().get("crash_count", 0) >= 1,
+            timeout=30.0,
+        ), "Expected guardian crash status to be published"
         assert _wait_until(lambda: continued_after_crash.is_set() or len(calls) >= 2), (
             "Expected guardian loop to continue probing"
         )
